+++ utils/nurl/url.go
@@ -1,4 +1,4 @@
-package url
+package nurl
 
 import (
 	"errors"
@@ -9,45 +9,8 @@ import (
 )
 
 type Parts struct {
-	Transport string //tcp
-	Host      string
-	Port      string
-}
-
-func SplitURL(url string) Parts {
-	var o Parts
-	u := strings.SplitN(url, "://", 2)
-	host := ""
-	if len(u) == 2 {
-		o.Transport = u[0]
-		host = u[1]
-	}
-	p := strings.Split(host, ":")
-	o.Host = p[0]
-	o.Port = p[1]
-	return o
-}
-
-func IsTCP(target string) bool {
-	if strings.HasPrefix(target, "tcp://") {
-		return true
-	} else {
-		return false
-	}
-}
-
-func JoinURL(u Parts) (url string, err error) {
-	t := u.Transport
-	h := u.Host
-	p := u.Port
-	if !ValidIP4(h) {
-		return "", errors.New("in valid url try ie: 192.168.1.1")
-	}
-	if !ValidPort(p) {
-		return "", errors.New("in valid url try ie: 8080 as a string")
-	}
-	ip := fmt.Sprintf("%s://%s:%s", t, h, p)
-	return ip, nil
+	Host string
+	Port string
 }
 
 func JoinIPPort(u Parts) (url string, err error) {
